Enhance the splash screen's Analytics section.
