+++ assets/js/components/setup/SetupUsingProxyWithSignIn/SplashContent.js
@@ -24,37 +24,22 @@ import PropTypes from 'prop-types';
 /**
  * WordPress dependencies
  */
-import {
-	createInterpolateElement,
-	useCallback,
-	Fragment,
-} from '@wordpress/element';
+import { createInterpolateElement, Fragment } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import Checkbox from '@/js/googlesitekit/components-gm2/Checkbox';
 import CompatibilityChecks from '@/js/components/setup/CompatibilityChecks';
 import Link from '@/js/components/Link';
 import P from '@/js/components/Typography/P';
-import Badge from '@/js/components/Badge';
 import ResetNotice from './ResetNotice';
 import SplashScreenshotSVG from './SetupFlowSVG';
+import AnalyticsOptIn from './AnalyticsOptIn';
 import SplashBackground from '@/svg/graphics/splash-graphic.svg';
 import Typography from '@/js/components/Typography';
-import useFormValue from '@/js/hooks/useFormValue';
-import useViewContext from '@/js/hooks/useViewContext';
-import {
-	ANALYTICS_NOTICE_CHECKBOX,
-	ANALYTICS_NOTICE_FORM_NAME,
-} from '@/js/components/setup/constants';
 import { Cell, Row } from '@/js/material-components';
-import { CORE_FORMS } from '@/js/googlesitekit/datastore/forms/constants';
 import { DISCONNECTED_REASON_CONNECTED_URL_MISMATCH } from '@/js/googlesitekit/datastore/user/constants';
-import { useDispatch, useSelect } from '@/js/googlesitekit-data';
-import { CORE_SITE } from '@/js/googlesitekit/datastore/site/constants';
-import { trackEvent } from '@/js/util';
 
 export default function SplashContent( {
 	analyticsModuleActive,
@@ -69,33 +54,9 @@ export default function SplashContent( {
 	showLearnMoreLink,
 	title,
 } ) {
-	const viewContext = useViewContext();
-
-	const { setValues } = useDispatch( CORE_FORMS );
-
-	const checked = useFormValue(
-		ANALYTICS_NOTICE_FORM_NAME,
-		ANALYTICS_NOTICE_CHECKBOX
-	);
-
-	const handleOnChange = useCallback(
-		( event ) => {
-			setValues( ANALYTICS_NOTICE_FORM_NAME, {
-				[ ANALYTICS_NOTICE_CHECKBOX ]: event.target.checked,
-			} );
-		},
-		[ setValues ]
-	);
-
 	const cellDetailsProp = analyticsModuleActive
 		? { smSize: 4, mdSize: 6, lgSize: 6 }
 		: { smSize: 4, mdSize: 7, lgSize: 6 };
-
-	const learnMoreLink = useSelect( ( select ) => {
-		return select( CORE_SITE ).getDocumentationLinkURL(
-			'setup-update-ga4-account'
-		);
-	} );
 
 	return (
 		<Fragment>
@@ -181,55 +142,7 @@ export default function SplashContent( {
 						) }
 
 					{ analyticsModuleAvailable && ! analyticsModuleActive && (
-						<div className="googlesitekit-setup__analytics-opt-in-wrapper">
-							<Checkbox
-								id="googlesitekit-analytics-setup-opt-in"
-								name="googlesitekit-analytics-setup-opt-in"
-								description={ createInterpolateElement(
-									__(
-										'To get better insights about your site, Site Kit will update your Analytics account, for example by enabling enhanced measurement. <LearnMoreLink /> <RecommendedBadge />',
-										'google-site-kit'
-									),
-									{
-										LearnMoreLink: (
-											<Link
-												href={ learnMoreLink }
-												onClick={ () => {
-													trackEvent(
-														viewContext,
-														'click_learn_more_link',
-														'analytics_checkbox'
-													);
-												} }
-												external
-											>
-												{ __(
-													'Learn more',
-													'google-site-kit'
-												) }
-											</Link>
-										),
-										RecommendedBadge: (
-											<Badge
-												className="googlesitekit-splash__analytics-recommended-badge"
-												label={ __(
-													'Recommended',
-													'google-site-kit'
-												) }
-											/>
-										),
-									}
-								) }
-								checked={ checked }
-								onChange={ handleOnChange }
-								value="1"
-							>
-								{ __(
-									'Get visitor insights by connecting Google Analytics as part of setup',
-									'google-site-kit'
-								) }
-							</Checkbox>
-						</div>
+						<AnalyticsOptIn />
 					) }
 
 					<CompatibilityChecks>{ children }</CompatibilityChecks>
